La funzione set_piece_color imposta un colore per ogni pezzo, prendendo come parametro il suo color_id. Viene chiamata sia in draw_active_piece che in draw_board, così che anche quando il pezzo viene fissato e diventa parte della board, mantiene il suo colore.

+++ game.h
@@ -2,6 +2,7 @@
 #define GAME_H
 #include <stdint.h>
 #include <stdbool.h>
+#include <SDL2/SDL.h>
 
 #define ROWS 20
 #define COLS 10
@@ -17,7 +18,7 @@ typedef struct{
 //struttura di un pezzo
 typedef struct {
     Offset shape[5]; //cambiato da 4 a 5 per realizzare dei pentamini al posto dei tetramini
-    int color_id;
+    uint8_t color_id;
 } Piece;
 
 //byte_field
